refactor(daemon): extract SkillDraftEventPublisher into skill/ package (batch A)
First batch of the AceClawDaemon decomposition. Lifts 5 fan-out methods
+ 3 pure helpers out of the 3,112-LoC god class into a dedicated
collaborator under a new dev.aceclaw.daemon.skill subpackage.

What moved:
- publishSkillDraftCreatedEvents / ValidationEvents / ReleaseEvents
  (rename -> publishCreated / publishValidationChanged / publishReleaseChanged
  since the class name already says "SkillDraft")
- toValidationJson / toReleaseJson  (RPC reply serializers)
- skillNameFromDraftPath / parseDraftFrontmatter / stripQuotes
  (pure parsing helpers, package-private for direct testing)

Why a subpackage:
The daemon's top-level package was already 70+ classes. Going forward
the decomposition will land in dev.aceclaw.daemon.{boot,wireup,scheduler,skill}
so each cluster's surface is browseable on its own. This batch is the
warmup that proves the boundary works before tackling wireAgentHandler
(1,632 LoC) and start()/shutdown().

Behavior preservation:
All 5 methods are byte-for-byte the same body, only relocated. 27 call
sites in AceClawDaemon migrate from local-method to skillDraftEventPublisher.
delegate. Constructor wired right after the recorders are built, before
anything else can use them.

New tests:
14 unit tests against the pure helpers in SkillDraftEventPublisherTest:
- skillNameFromDraftPath: 4 cases (parent dir, Windows \\, root file, nested)
- stripQuotes: 5 cases (double, single, mismatched, unquoted, null/short)
- parseDraftFrontmatter: 5 cases (happy path, key lowercase, missing, unclosed,
  colon-less lines skipped)
The publish methods themselves are covered by existing daemon end-to-end
integration tests.

LoC:
- AceClawDaemon: 3,112 -> 2,909 (-203, -6.5%)
- New SkillDraftEventPublisher: 263 LoC
- New SkillDraftEventPublisherTest: 142 LoC
